feature: Add steady state pattern probabilities
Using linear algebra, calculate the steady state probability of the
PROBABILITY_MATRIX Markov chain to provide the chances of each pattern
occurring when the user does not know the previous weeks' pattern.

Calculations are done on the fly, at the cost of a little extra time, to
make probability distribution updates easier for future developers.  If
the chances are updated (because of an update or for a different game in
the series), no additional knowledge is required of the developer to
provide these steady states.

Closes: #68