Adds classnames to weekly min/max table cells
Prior to this commit, `scripts.js` did not place class names on any of
the cells in the predictions table. This commit adds `.lowest` and
`.highest` classes to cells in the table using the following logic:

* `.lowest` is applied to any cell where the low end of that day's
    range matches the week's lowest guaranteed minimum value.
* `.highest` is applied to any cell where the high end of that day's
    range matches the week's highest possible maximum value.
* If `day.min === day.max`, `.highest` takes precedence over `.lowest`.

Additionally, all of a day's cells now feature `[data-min]` and
`[data-max]` values. If `day.min === day.max`, both data attributes
will contain the same value.

These classes are presently unstyled, and these data attributes are
currently unused, but both of these changes seek to:

* Enable easier styling and simpler feature development in the future
* Make it easier for future developers to find these cells to make
    modifications in future PRs.