fix: Allow people to input the buy price even as a first time buyer
This will give us a basis for colouring. Also remove the Sunday prices from the table, to avoid confusion and questions when people ask why the Sunday price is different to the one they put in. This has the added bonus of making the table smaller, and doesn't really remove any relevant information for the future.

Fixes #288