Compute "Expected value" data in the chart and at the bottom of the table
The expected value is the weighted average of each column of regular data
in the table. My island is roughly 50/50 large spike vs decreasing
today, and without an expected value it's difficult to tell whether to
sell now to cut my losses or wait until later in the week in hopes that
it's the large spike.