Fixed some sentences
In spanish, we have two possessive adjectives, "tu" and "su". The first one is used in a informal conversation, the second one in a formal conversation. All the text is based in a informal conversation, but in line 28, there is the "su" possessive adjective.

In line 72, "el" cannot be added, as the noun "Turnip Prophet" is a proper noun.